requirements.txt
requests==2.31.0
httpx==0.27.0
twikit==2.1.2
python-dotenv==1.0.1